requirements.txt
joblib==1.2.0
logzero==1.7.0
numpy==1.24.2
open-clip-torch==2.16.0
pandas==2.0.0
Pillow==9.4.0
scipy==1.10.1
sentence-transformers==2.2.2
tensorboard==2.12.0
timm==0.8.19.dev0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
